# swagger.yaml
openapi: "3.0.0"
paths:
...
  /profile/find-by-tags:
    get:
      operationId: "ProfileController_findByTag"
      parameters:
        - name: "tags"
          required: true
          in: "query"
          schema:
            type: "array"
            items:
              type: "string"
      responses:
        200:
          description: "Perfiles obtenidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindByTagsProfileResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /profile/find-by-tag-groups:
    get:
      operationId: "ProfileController_findByTagGroups"
      parameters:
        - name: "tagGroups"
          required: true
          in: "query"
          schema:
            type: "array"
            items:
              type: "string"
      responses:
        200:
          description: "Perfiles obtenidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindByTagGroupsProfileResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /profile/find-by-date-range:
    get:
      operationId: "ProfileController_findByDateRange"
      parameters:
        - name: "from"
          required: true
          in: "query"
          schema:
            type: "string"
        - name: "to"
          required: true
          in: "query"
          schema:
            type: "string"
      responses:
        200:
          description: "Perfiles obtenidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindByDateRangeResponseDto"
        412:
          description: "Fecha inválida"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /profile/find-by-phone-number/{phoneNumber}:
    get:
      operationId: "ProfileController_findByPhoneNumber"
      parameters:
        - name: "phoneNumber"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Perfil obtenido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindByPhoneNumberResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
  /profile/create:
    post:
      operationId: "ProfileController_create"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileDto"
      responses:
        200:
          description: "Perfil creado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProfileResponseDto"
        404:
          description: "No se encontró la etiqueta de participante"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        409:
          description: "Error en la creación del perfil"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /profile/{id}:
    get:
      operationId: "ProfileController_findById"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Perfil obtenido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindByIdProfileResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      operationId: "ProfileController_delete"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Perfil eliminado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteProfileResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    patch:
      operationId: "ProfileController_update"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileDto"
      responses:
        200:
          description: "Perfil actualizado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProfileDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        409:
          description: "Error en la actualización del perfil"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
info:
  title: "Expo Backend"
  description: "Backend de las aplicaciones de Expo"
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    ErrorDto:
      type:
        - "object"
      properties:
        message:
          type:
            - "array"
          items:
            type:
              - "string"
        statusCode:
          type:
            - "number"
        error:
          type:
            - "string"
      required:
        - "message"
        - "statusCode"
        - "error"
...
    FindByTagsProfileResponseDto:
      type:
        - "object"
      properties:
        profiles:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              id:
                type:
                  - "string"
                format: "uuid"
              shortId:
                type:
                  - "number"
              phoneNumber:
                type:
                  - "string"
                minLength: 1
              secondaryPhoneNumber:
                type:
                  - "string"
                  - "null"
              fullName:
                type:
                  - "string"
                minLength: 1
              firstName:
                type:
                  - "string"
                  - "null"
                minLength: 1
              gender:
                type:
                  - "string"
                  - "null"
              birthDate:
                type:
                  - "string"
                  - "null"
              profilePictureUrl:
                type:
                  - "string"
                  - "null"
                format: "uri"
              instagram:
                type:
                  - "string"
                  - "null"
              mail:
                type:
                  - "string"
                  - "null"
                format: "email"
              dni:
                type:
                  - "string"
                  - "null"
              alternativeNames:
                type:
                  - "array"
                items:
                  type:
                    - "string"
              birthLocationId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              residenceLocationId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              isInTrash:
                type:
                  - "boolean"
              movedToTrashDate:
                type:
                  - "string"
                  - "null"
                format: "date-time"
              created_at:
                type:
                  - "string"
                format: "date-time"
              updated_at:
                type:
                  - "string"
                format: "date-time"
              tags:
                type:
                  - "array"
                items:
                  type:
                    - "object"
                  properties:
                    id:
                      type:
                        - "string"
                      format: "uuid"
                    name:
                      type:
                        - "string"
                      minLength: 1
                    groupId:
                      type:
                        - "string"
                      format: "uuid"
                    type:
                      type:
                        - "string"
                      enum:
                        - "PROFILE"
                        - "EVENT"
                        - "PARTICIPANT"
                        - "NOT_IN_SYSTEM"
                    created_at:
                      type:
                        - "string"
                      format: "date-time"
                    updated_at:
                      type:
                        - "string"
                      format: "date-time"
                    group:
                      type:
                        - "object"
                      properties:
                        isExclusive:
                          type:
                            - "boolean"
                      required:
                        - "isExclusive"
                  required:
                    - "id"
                    - "name"
                    - "groupId"
                    - "type"
                    - "created_at"
                    - "updated_at"
                    - "group"
            required:
              - "id"
              - "shortId"
              - "phoneNumber"
              - "secondaryPhoneNumber"
              - "fullName"
              - "firstName"
              - "gender"
              - "birthDate"
              - "profilePictureUrl"
              - "instagram"
              - "mail"
              - "dni"
              - "alternativeNames"
              - "birthLocationId"
              - "residenceLocationId"
              - "isInTrash"
              - "movedToTrashDate"
              - "created_at"
              - "updated_at"
              - "tags"
      required:
        - "profiles"
    FindByTagGroupsProfileResponseDto:
      type:
        - "object"
      properties:
        profiles:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              id:
                type:
                  - "string"
                format: "uuid"
              shortId:
                type:
                  - "number"
              phoneNumber:
                type:
                  - "string"
                minLength: 1
              secondaryPhoneNumber:
                type:
                  - "string"
                  - "null"
              fullName:
                type:
                  - "string"
                minLength: 1
              firstName:
                type:
                  - "string"
                  - "null"
                minLength: 1
              gender:
                type:
                  - "string"
                  - "null"
              birthDate:
                type:
                  - "string"
                  - "null"
              profilePictureUrl:
                type:
                  - "string"
                  - "null"
                format: "uri"
              instagram:
                type:
                  - "string"
                  - "null"
              mail:
                type:
                  - "string"
                  - "null"
                format: "email"
              dni:
                type:
                  - "string"
                  - "null"
              alternativeNames:
                type:
                  - "array"
                items:
                  type:
                    - "string"
              birthLocationId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              residenceLocationId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              isInTrash:
                type:
                  - "boolean"
              movedToTrashDate:
                type:
                  - "string"
                  - "null"
                format: "date-time"
              created_at:
                type:
                  - "string"
                format: "date-time"
              updated_at:
                type:
                  - "string"
                format: "date-time"
              tags:
                type:
                  - "array"
                items:
                  type:
                    - "object"
                  properties:
                    id:
                      type:
                        - "string"
                      format: "uuid"
                    name:
                      type:
                        - "string"
                      minLength: 1
                    groupId:
                      type:
                        - "string"
                      format: "uuid"
                    type:
                      type:
                        - "string"
                      enum:
                        - "PROFILE"
                        - "EVENT"
                        - "PARTICIPANT"
                        - "NOT_IN_SYSTEM"
                    created_at:
                      type:
                        - "string"
                      format: "date-time"
                    updated_at:
                      type:
                        - "string"
                      format: "date-time"
                  required:
                    - "id"
                    - "name"
                    - "groupId"
                    - "type"
                    - "created_at"
                    - "updated_at"
            required:
              - "id"
              - "shortId"
              - "phoneNumber"
              - "secondaryPhoneNumber"
              - "fullName"
              - "firstName"
              - "gender"
              - "birthDate"
              - "profilePictureUrl"
              - "instagram"
              - "mail"
              - "dni"
              - "alternativeNames"
              - "birthLocationId"
              - "residenceLocationId"
              - "isInTrash"
              - "movedToTrashDate"
              - "created_at"
              - "updated_at"
              - "tags"
      required:
        - "profiles"
    FindByDateRangeResponseDto:
      type:
        - "object"
      additionalProperties:
        type:
          - "array"
        items:
          type:
            - "object"
          properties:
            id:
              type:
                - "string"
              format: "uuid"
            shortId:
              type:
                - "number"
            phoneNumber:
              type:
                - "string"
              minLength: 1
            secondaryPhoneNumber:
              type:
                - "string"
                - "null"
            fullName:
              type:
                - "string"
              minLength: 1
            firstName:
              type:
                - "string"
                - "null"
              minLength: 1
            gender:
              type:
                - "string"
                - "null"
            birthDate:
              type:
                - "string"
                - "null"
            profilePictureUrl:
              type:
                - "string"
                - "null"
              format: "uri"
            instagram:
              type:
                - "string"
                - "null"
            mail:
              type:
                - "string"
                - "null"
              format: "email"
            dni:
              type:
                - "string"
                - "null"
            alternativeNames:
              type:
                - "array"
              items:
                type:
                  - "string"
            birthLocationId:
              type:
                - "string"
                - "null"
              format: "uuid"
            residenceLocationId:
              type:
                - "string"
                - "null"
              format: "uuid"
            isInTrash:
              type:
                - "boolean"
            movedToTrashDate:
              type:
                - "string"
                - "null"
              format: "date-time"
            created_at:
              type:
                - "string"
              format: "date-time"
            updated_at:
              type:
                - "string"
              format: "date-time"
            tags:
              type:
                - "array"
              items:
                type:
                  - "object"
                properties:
                  id:
                    type:
                      - "string"
                    format: "uuid"
                  name:
                    type:
                      - "string"
                    minLength: 1
                  groupId:
                    type:
                      - "string"
                    format: "uuid"
                  type:
                    type:
                      - "string"
                    enum:
                      - "PROFILE"
                      - "EVENT"
                      - "PARTICIPANT"
                      - "NOT_IN_SYSTEM"
                  created_at:
                    type:
                      - "string"
                    format: "date-time"
                  updated_at:
                    type:
                      - "string"
                    format: "date-time"
                  group:
                    type:
                      - "object"
                    properties:
                      id:
                        type:
                          - "string"
                        format: "uuid"
                    required:
                      - "id"
                required:
                  - "id"
                  - "name"
                  - "groupId"
                  - "type"
                  - "created_at"
                  - "updated_at"
                  - "group"
          required:
            - "id"
            - "shortId"
            - "phoneNumber"
            - "secondaryPhoneNumber"
            - "fullName"
            - "firstName"
            - "gender"
            - "birthDate"
            - "profilePictureUrl"
            - "instagram"
            - "mail"
            - "dni"
            - "alternativeNames"
            - "birthLocationId"
            - "residenceLocationId"
            - "isInTrash"
            - "movedToTrashDate"
            - "created_at"
            - "updated_at"
            - "tags"
...